+++ src/java/models/Cart.java
@@ -1,5 +1,6 @@
 package models;
 
+import java.math.BigInteger;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +9,7 @@ import java.sql.Timestamp;
 public class Cart {
 
     private int cartId;
-    private int totalPrice;
+    private BigInteger totalPrice;
     private Timestamp payDay;
     private String status;
     private Date startDate;
@@ -27,7 +28,7 @@ public class Cart {
     public Cart() {
     }
 
-    public Cart(int cartId, int totalPrice, Timestamp payDay, String status, Date startDate, Date endDate, boolean isActive, boolean isPayment, Customer customer, PaymentMethod paymentMethod) {
+    public Cart(int cartId, BigInteger totalPrice, Timestamp payDay, String status, Date startDate, Date endDate, boolean isActive, boolean isPayment, Customer customer, PaymentMethod paymentMethod) {
         this.cartId = cartId;
         this.totalPrice = totalPrice;
         this.payDay = payDay;
@@ -56,11 +57,11 @@ public class Cart {
         this.cartId = cartId;
     }
 
-    public int getTotalPrice() {
+    public BigInteger getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigInteger totalPrice) {
         this.totalPrice = totalPrice;
     }
